refactor: promote accessibility, hud, and mainthread to pkg/
These are self-contained, general-purpose macOS utilities with no domain or app
dependencies — Accessibility permission checks, a system-bezel-style HUD overlay,
and main-thread (GCD) dispatch — so move them from internal/infra to pkg/ next to
inputsource, where they are usable as standalone libraries. keyboard stays
internal as it is specific to the Globe key and reverse-modifier behavior.
